refactor: front-end code
