+++ hr_governance/static/src/components/governance_form.xml
@@ -1,30 +1,10 @@
 <?xml version="1.0" encoding="UTF-8" ?>
 <templates xml:space="preserve">
-
     <t
-        t-name="hr_governance.FormWrapperView"
+        t-name="hr_governance.GovernanceFormView"
         t-inherit="web.FormView"
         t-inherit-mode="primary"
     >
-        <xpath
-            expr="//t[@t-set-slot='control-panel-additional-actions']"
-            position="replace"
-        >
-            <t t-set-slot="control-panel-additional-actions">
-                <t t-if="canEdit">
-                    <CogMenu
-                        getActiveIds="() => model.root.isNew ? [] : [model.root.resId]"
-                        context="props.context"
-                        items="props.info.actionMenus ? this.actionMenuItems : {}"
-                        isDomainSelected="model.root.isDomainSelected"
-                        resModel="model.root.resModel"
-                        domain="props.domain"
-                        onActionExecuted="() => model.load({ resId: model.root.resId, resIds: model.root.resIds })"
-                        shouldExecuteAction.bind="shouldExecuteAction"
-                    />
-                </t>
-            </t>
-        </xpath>
         <xpath
             expr="//t[@t-set-slot='control-panel-create-button'][2]"
             position="replace"
