--- hr_governance/static/src/views/circlepack_controller.xml
@@ -1,21 +1,0 @@
-<?xml version="1.0" encoding="utf-8" ?>
-<templates>
-    <t t-name="hr_governance.CirclePackView">
-        <div
-            t-attf-class="{ isMobile ? 'o_action_delegate_scroll' : '' }} {{ props.className }}"
-        >
-            <Layout
-                className="(model.useSampleModel ? 'o_view_sample_data' : '') + ' d-flex'"
-                display="props.display"
-            >
-                <t t-set-slot="layout-actions">
-                    <Many2ManySearchBar />
-                </t>
-                <t t-if="model.data">
-                    <t t-component="props.Renderer" records="model.data" />
-                </t>
-                <t t-else="" t-call="web.NoContentHelper" />
-            </Layout>
-        </div>
-    </t>
-</templates>
