--- hr_governance/static/src/views/circlepack_renderer.xml
@@ -1,26 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" ?>
-<templates>
-    <t t-name="hr_governance.CirclePackRenderer">
-        <div class="o_circle_packing_container" t-ref="container">
-            <div
-                class="o_circle_packing_svg_container"
-                t-ref="chart"
-                style="flex: 1 1 100%"
-            />
-            <t t-if="state.activeResId">
-                <div class="chart-splitter" t-on-mousedown="startResizing" />
-                <div class="form_wrapper" style="flex: 0 1 0%">
-                    <View
-                        type="'form'"
-                        resModel="'governance.circle'"
-                        resId="state.activeResId"
-                        t-key="state.activeResId"
-                        loadActionMenus="true"
-                        preventCreate="preventCreate"
-                        preventEdit="preventEdit"
-                    />
-                </div>
-            </t>
-        </div>
-    </t>
-</templates>
